requirements.txt
Django==3.1.7
djangorestframework==3.12.2
model-bakery==1.2.1
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-django==4.1.0
